minor attrs changes + adding new viz module for plots
